Broke up IRender monolith and moved it into its separate classes
