+++ src/EPPlus.DrawingRenderer/Svg/Core/SvgBaseRenderer.cs
