+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgEllipseRenderer.cs
