+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgGroupRenderer.cs
