+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgLineRenderer.cs
