+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgParagraphRenderer.cs
