+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgPathRenderer.cs
