+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgRectRenderer.cs
