+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgTextRenderer.cs
