+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgTextRunRenderer.cs
