+++ src/EPPlus.DrawingRenderer/Svg/ItemRenderers/SvgTitleRenderer.cs
