+++ src/EPPlus/Drawing/Renderer/Chart/ChartLegendRenderer.cs
@@ -636,29 +636,29 @@ namespace EPPlusImageRenderer.Svg
             //adjust datapoints if not enough
             if(catValues.Count > ps.DataPoints.Count)
             {
-                var prevPointCount = ps.DataPoints.Count;
-                ps.DataPoints.ClearDataPoints();
-                var pieChart = ct.As.Chart.PieChart;
+                //var prevPointCount = ps.DataPoints.Count;
+                //ps.DataPoints.ClearDataPoints();
+                //var pieChart = ct.As.Chart.PieChart;
 
-                for (int i = 0; i< catValues.Count; i++)
-                {
-                    ps.DataPoints.Add(i);
-                    var defaultFill = DefaultFillColor;
+                //for (int i = 0; i< catValues.Count; i++)
+                //{
+                //    ps.DataPoints.Add(i);
+                //    var defaultFill = DefaultFillColor;
 
-                    if (ct.VaryColors)
-                    {
-                        if (ct.StyleManager.Style == null)
-                        {
-                            var mod5 = i % 5;
-                            //TODO: Only works for base-case. Add support for patterns 1,3 and 4 instead of just 2 as basecase
-                            defaultFill = ChartRenderer.Theme.ColorScheme.GetColorByEnum(OfficeOpenXml.Drawing.eSchemeColor.Accent1 + mod5).GetColor();
-                        }
-                    }
-                    if(defaultFill != null && i > prevPointCount-1)
-                    {
-                        ps.DataPoints[i].Fill.Color = defaultFill.Value;
-                    }
-                }
+                //    if (ct.VaryColors)
+                //    {
+                //        if (ct.StyleManager.Style == null)
+                //        {
+                //            var mod5 = i % 5;
+                //            //TODO: Only works for base-case. Add support for patterns 1,3 and 4 instead of just 2 as basecase
+                //            defaultFill = ChartRenderer.Theme.ColorScheme.GetColorByEnum(OfficeOpenXml.Drawing.eSchemeColor.Accent1 + mod5).GetColor();
+                //        }
+                //    }
+                //    if(defaultFill != null && i > prevPointCount-1)
+                //    {
+                //        ps.DataPoints[i].Fill.Color = defaultFill.Value;
+                //    }
+                //}
             }
 
             double lastWidth = 0;
@@ -697,6 +697,7 @@ namespace EPPlusImageRenderer.Svg
                 lastWidth = tbWidth + si.Width;
 
                 totalWidth += tbWidth + si.Width + maxIconLength + MarginIconText;
+
 
                 var dp = ps.DataPoints[i];
